Avoid 'conversion may change value' cases in sizecalculator.cpp and filesystem.cpp
Revision: https://invent.kde.org/utilities/krusader/-/merge_requests/31

+++ krusader/FileSystem/filesystem.cpp
@@ -268,7 +268,7 @@ QString FileSystem::readLinkSafely(const char *path)
 
         // read success
         if (nBytesRead < bufferSize || nBytesRead == maxBufferSize) {
-            return QString::fromLocal8Bit(buffer.get(), nBytesRead);
+            return QString::fromLocal8Bit(buffer.get(), static_cast<int>(nBytesRead));
         }
 
         // increase the buffer and retry again
